+++ tests/bootstrap.php
@@ -13,17 +13,15 @@ if ( ! file_exists( $_tests_dir . '/includes/functions.php' ) ) {
 }
 
 // Give access to tests_add_filter() function.
-require_once 'includes/functions.php';
+require_once $_tests_dir . '/includes/functions.php';
 
 /**
  * Manually load the plugin being tested.
  */
 function _manually_load_plugin() {
-	$_tests_plugin = getenv( 'WP_PLUGIN_FILE' );
-	// require dirname( dirname( __FILE__ ) ) . '/' . $_tests_plugin;.
 	require dirname( dirname( __FILE__ ) ) . '/block-lab.php';
 }
 tests_add_filter( 'muplugins_loaded', '_manually_load_plugin' );
 
 // Start up the WP testing environment.
-require 'includes/bootstrap.php';
+require $_tests_dir . '/includes/bootstrap.php';